MueLu: NodeComm generation seems to work
