+++ packages/muelu/src/Utils/MueLu_UtilitiesBase_decl.hpp
@@ -84,13 +84,6 @@
 #include "MueLu_Exceptions.hpp"
 
 
-#ifdef HAVE_MPI
-#include <mpi.h>
-#include <netdb.h>
-#include <arpa/inet.h>
-#endif
-
-
 namespace MueLu {
 
 // MPI helpers
@@ -979,44 +972,6 @@ namespace MueLu {
     }
 
   }; // class Utils
-
-
-    // Generates a communicator whose only members are other ranks of the baseComm on my node
-    Teuchos::RCP<Teuchos::Comm<int> > GenerateNodeComm(const RCP<Teuchos::Comm<int> > & baseComm, int &NodeId) {
-#ifdef HAVE_MPI
-       int numRanks = baseComm->getRank();       
-       if(numRanks == 1) {NodeId = baseComm->getRank(); return baseComm;}
-
-       // Get an integer from the hostname
-       char hostname[MPI_MAX_PROCESSOR_NAME];
-       int len;
-       MPI_Get_processor_name(hostname,&len);
-       struct hostent * host = gethostbyname(hostname);
-       int myaddr = (int) htonl(inet_network(inet_ntoa(*(struct in_addr *)host->h_addr)));
-       
-       // All-to-all exchange of address integers
-       std::vector<int> addressList(numRanks);
-       Teuchos::gatherAll(*baseComm,1,&myaddr,numRanks,&addressList[0]);
-
-       // Sort!
-       std::sort(addressList.begin(),addressList.end());
-
-       // Find the number of node-buddies I have
-       int numNodes = 0;
-       for(int i=0, prev=addressList[0]; i<numRanks && prev != myaddr; i++) {
-         if(prev != addressList[i]) {
-           prev = addressList[i];
-           numNodes++;
-         }
-       }
-       NodeId = numNodes;
-       // Generate nodal communicator
-       return baseComm->split(NodeId,baseComm->getRank());   
-#else
-       NodeId = baseComm->getRank();
-       return baseComm;
-#endif
-    }
 
 
   ///////////////////////////////////////////
